Introduce ClusterBackend option to zeekctl.cfg and plugins
This adds a new global ClusterBackend option and adds 3 new plugins.
One for Broker, one for ZeroMQ, and one that ensures that the new
internal ClusterTopicSeparator setting is set by a cluster backend
and otherwise vetoes the install command with a descriptive message.

The separator is used by the WebSocket code to construct node specific
topics to publish to. For ZeroMQ this isn't exactly important as topics
are just a sequence of bytes (same for Broker), but for other pub-sub
systems, wild-card or prefix matching may only be possible on parts of
the topic (e.g., NATS subject has a more rigid structure).